Wrap the workflow enforcer in a class
We want to be able to globally change the behaviour of the workflow
enforcer. Turning it into a class means we can use class variables to
do that later.
